test(composition): partition exact result verification
